move openapi_doc.py
